Refactoring mockAPI.js
