--- src/data/mockData.js
@@ -1,9 +1,0 @@
-export const BrandName = 'futurspace';
-
-export const menus = [
-  { menu: 'about', href: '/about' },
-  { menu: 'pricing', href: '/pricing' },
-  { menu: 'offices', href: '/offices' },
-  { menu: 'spaces', href: '/spaces' },
-  { menu: 'contact us', href: '/contact' },
-];
