+++ src/utils/mockAPI.js
@@ -174,3 +174,13 @@ export const faqsList = [
     answer: 'Lorem ipsum dolor sit amet consectetur adipiscing elit elementum velit nunc tortor pulvinar ornare',
   },
 ];
+
+export const BrandName = 'futurspace';
+
+export const menus = [
+  { menu: 'about', href: '/about' },
+  { menu: 'pricing', href: '/pricing' },
+  { menu: 'offices', href: '/offices' },
+  { menu: 'spaces', href: '/spaces' },
+  { menu: 'contact us', href: '/contact' },
+];
